Move HashableValue to client module
The `HashableValue` type is a private type used only internally in the
`client` module, no need for a separate module to hold it.

Move `HashableValue` to the `client` module and make it private.
